LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-compact-buttons-styles-expected.txt

Testing the MacOSInlineMediaControls computed styles in compact mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PlayPauseButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "6px"
PASS getComputedStyle(button.element).height is "25px"

SkipBackButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "6px"
PASS getComputedStyle(button.element).height is "25px"

MuteButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "6px"
PASS getComputedStyle(button.element).height is "25px"

FullscreenButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "6.5px"
PASS getComputedStyle(button.element).height is "25px"
PASS successfullyParsed is true

TEST COMPLETE
